Check the google drive link

+++ main.py
@@ -10,18 +10,21 @@ def download_single_file(url, destination):
 
 
 def download_a_folder(url, destination):
-    file_list = gdown.download_folder(url)
+    if 'drive/folders' in url:
+        file_list = gdown.download_folder(url)
 
-    # Move download files to given place
-    for file in file_list:
-        if not os.path.isdir(file):
-            shutil.move(file, destination)
+        # Move download files to given place
+        for file in file_list:
+            if not os.path.isdir(file):
+                shutil.move(file, destination)
 
-    # Delete the source folder
-    for file in file_list:
-        if os.path.exists(Path(file).parent):
-            shutil.rmtree(Path(file).parent)
-            print(f"Remove folder: {str(Path(file).parent)}")
+        # Delete the source folder
+        for file in file_list:
+            if os.path.exists(Path(file).parent):
+                shutil.rmtree(Path(file).parent)
+                print(f"Remove folder: {str(Path(file).parent)}")
+    else:
+        print('Please enter a valid url')
 
 if __name__ == '__main__':
     # Handle a folder
